bug(ci): floating stable Rust toolchain breaks clippy gate on pre-existing, unrelated code (#3539)
* fix(ci): pin Rust toolchain and fix pre-existing clippy violations

A floating `stable` toolchain in CI picked up new/tightened clippy
lints (useless_format, needless_late_init) between runs, flipping
previously-green, untouched files to failing for every open PR
regardless of its actual diff. Fix the two flagged lints and pin
the toolchain to 1.94.1 so future upgrades are a deliberate,
reviewed bump instead of a silent CI break.


* fix(tests): remove unnecessary i32 casts flagged by pinned clippy

store::increment() already returns i32; the trailing `as i32` on both
codex/kimi attempt bindings in the review-invocations test tripped
clippy's unnecessary_cast lint under the toolchain pin from cbe47784.


---------

+++ src/store/tests.rs
@@ -5222,7 +5222,7 @@ async fn review_invocations_preserved_across_reset_prevents_task_run_overwrite()
     let codex_attempt = store
         .increment(task_id, "review_invocations")
         .await
-        .unwrap() as i32; // → 2
+        .unwrap(); // → 2
     let codex_run_id = store
         .start_run(&StartRun {
             task_id,
@@ -5253,7 +5253,7 @@ async fn review_invocations_preserved_across_reset_prevents_task_run_overwrite()
     let kimi_attempt = store
         .increment(task_id, "review_invocations")
         .await
-        .unwrap() as i32; // → 3
+        .unwrap(); // → 3
     let kimi_run_id = store
         .start_run(&StartRun {
             task_id,
